+++ plugins/org.preesm.codegen/src/org/preesm/codegen/model/generator/CodegenModelGenerator.java
@@ -111,6 +111,7 @@ import org.preesm.codegen.model.CoreBlock;
 import org.preesm.codegen.model.DataTransferAction;
 import org.preesm.codegen.model.Delimiter;
 import org.preesm.codegen.model.Direction;
+import org.preesm.codegen.model.DistributedBuffer;
 import org.preesm.codegen.model.DistributedMemoryCommunication;
 import org.preesm.codegen.model.FifoCall;
 import org.preesm.codegen.model.FifoOperation;
@@ -130,7 +131,6 @@ import org.preesm.codegen.model.SharedMemoryCommunication;
 import org.preesm.codegen.model.SpecialCall;
 import org.preesm.codegen.model.SpecialType;
 import org.preesm.codegen.model.SubBuffer;
-import org.preesm.codegen.model.TwinBuffer;
 import org.preesm.codegen.model.Variable;
 import org.preesm.codegen.model.util.CodegenModelUserFactory;
 import org.preesm.commons.exceptions.PreesmException;
@@ -471,7 +471,7 @@ public class CodegenModelGenerator extends AbstractCodegenModelGenerator {
           break;
 
         case VertexType.TYPE_SEND:
-          if (buffer instanceof TwinBuffer) {
+          if (buffer instanceof DistributedBuffer) {
             generateDistributedCommunication(operatorBlock, vert, VertexType.TYPE_SEND);
           } else {
             generateCommunication(operatorBlock, vert, VertexType.TYPE_SEND);
@@ -479,7 +479,7 @@ public class CodegenModelGenerator extends AbstractCodegenModelGenerator {
           break;
 
         case VertexType.TYPE_RECEIVE:
-          if (buffer instanceof TwinBuffer) {
+          if (buffer instanceof DistributedBuffer) {
             generateDistributedCommunication(operatorBlock, vert, VertexType.TYPE_RECEIVE);
           } else {
             generateCommunication(operatorBlock, vert, VertexType.TYPE_RECEIVE);
@@ -949,7 +949,8 @@ public class CodegenModelGenerator extends AbstractCodegenModelGenerator {
 
             // also accessible with dagAlloc.getKey().getWeight()
             dagEdgeBuffer.setSize(dagEdgeSize);
-            TwinBuffer duplicatedBuffer = generateTwinBuffer(this.dagEdgeBuffers.get(originalDagEdge), dagEdgeBuffer);
+            DistributedBuffer duplicatedBuffer = generateDistributedBuffer(this.dagEdgeBuffers.get(originalDagEdge),
+                dagEdgeBuffer);
             this.dagEdgeBuffers.put(originalDagEdge, duplicatedBuffer);
             /**
              * Notes for future development of this feature If you are reading this because you want to adapt the code
@@ -1138,22 +1139,16 @@ public class CodegenModelGenerator extends AbstractCodegenModelGenerator {
       // Get the corresponding Variable
       final Variable varFirstFound = this.srSDFEdgeBuffers.get(subBufferProperties);
       Variable var = null;
-      if (varFirstFound instanceof TwinBuffer) {
-        TwinBuffer twinBuffer = (TwinBuffer) varFirstFound;
-        SubBuffer original = (SubBuffer) twinBuffer.getOriginal();
-        EList<Buffer> twins = twinBuffer.getTwins();
-        SubBuffer originalContainer = (SubBuffer) original.getContainer();
+      if (varFirstFound instanceof DistributedBuffer) {
+        DistributedBuffer distributedBuffer = (DistributedBuffer) varFirstFound;
+        EList<Buffer> repeatedBuffers = distributedBuffer.getDistributedCopies();
         String coreBlockName = dagVertex.getPropertyStringValue("Operator");
-        if (originalContainer.getContainer().getName().equals(coreBlockName)) {
-          var = original;
-        } else {
-          for (Buffer bufferTwinChecker : twins) {
-            SubBuffer subBufferChecker = (SubBuffer) bufferTwinChecker;
-            SubBuffer twinContainer = (SubBuffer) subBufferChecker.getContainer();
-            if (twinContainer.getContainer().getName().equals(coreBlockName)) {
-              var = subBufferChecker;
-              break;
-            }
+        for (Buffer bufferRepeatedChecker : repeatedBuffers) {
+          SubBuffer subBufferChecker = (SubBuffer) bufferRepeatedChecker;
+          SubBuffer repeatedContainer = (SubBuffer) subBufferChecker.getContainer();
+          if (repeatedContainer.getContainer().getName().equals(coreBlockName)) {
+            var = subBufferChecker;
+            break;
           }
         }
       } else {
@@ -1293,21 +1288,15 @@ public class CodegenModelGenerator extends AbstractCodegenModelGenerator {
       final Buffer firstFound = this.srSDFEdgeBuffers.get(subBufferProperties);
       Buffer buffer = null;
       String coreBlockName = dagVertex.getPropertyStringValue("Operator");
-      if (firstFound instanceof TwinBuffer) {
-        TwinBuffer twinBuffer = (TwinBuffer) firstFound;
-        SubBuffer original = (SubBuffer) twinBuffer.getOriginal();
-        SubBuffer originalContainer = (SubBuffer) original.getContainer();
-        EList<Buffer> twins = twinBuffer.getTwins();
-        if (originalContainer.getContainer().getName().equals(coreBlockName)) {
-          buffer = original;
-        } else {
-          for (Buffer twinChecker : twins) {
-            SubBuffer twinSubBuffer = (SubBuffer) twinChecker;
-            SubBuffer twinContainer = (SubBuffer) twinSubBuffer.getContainer();
-            if (twinContainer.getContainer().getName().equals(coreBlockName)) {
-              buffer = twinSubBuffer;
-              break;
-            }
+      if (firstFound instanceof DistributedBuffer) {
+        DistributedBuffer distributedBuffer = (DistributedBuffer) firstFound;
+        EList<Buffer> repeatedBuffers = distributedBuffer.getDistributedCopies();
+        for (Buffer repeatedBufferChecker : repeatedBuffers) {
+          SubBuffer repeatedSubBuffer = (SubBuffer) repeatedBufferChecker;
+          SubBuffer repeatedContainer = (SubBuffer) repeatedSubBuffer.getContainer();
+          if (repeatedContainer.getContainer().getName().equals(coreBlockName)) {
+            buffer = repeatedSubBuffer;
+            break;
           }
         }
       } else {
@@ -1511,21 +1500,16 @@ public class CodegenModelGenerator extends AbstractCodegenModelGenerator {
     // Find the corresponding DAGEdge buffer(s)
     final DAGEdge dagEdge = dagVertex.getPropertyBean()
         .getValue(ImplementationPropertyNames.SendReceive_correspondingDagEdge);
-    final TwinBuffer twinBuffer = (TwinBuffer) this.dagEdgeBuffers.get(dagEdge);
+    final DistributedBuffer distributedBuffer = (DistributedBuffer) this.dagEdgeBuffers.get(dagEdge);
     final String coreBlockName = operatorBlock.getName();
 
-    final SubBuffer original = (SubBuffer) twinBuffer.getOriginal();
-    final List<Buffer> twins = twinBuffer.getTwins();
+    final List<Buffer> repeatedBuffers = distributedBuffer.getDistributedCopies();
     Buffer buffer = null;
-    if (original.getContainer().getName().equals(coreBlockName)) {
-      buffer = original;
-    } else {
-      for (Buffer bufferTwinChecker : twins) {
-        SubBuffer subBufferChecker = (SubBuffer) bufferTwinChecker;
-        if (subBufferChecker.getContainer().getName().equals(coreBlockName)) {
-          buffer = subBufferChecker;
-          break;
-        }
+    for (Buffer bufferRepeatedChecker : repeatedBuffers) {
+      SubBuffer subBufferChecker = (SubBuffer) bufferRepeatedChecker;
+      if (subBufferChecker.getContainer().getName().equals(coreBlockName)) {
+        buffer = subBufferChecker;
+        break;
       }
     }
     if (buffer == null) {
@@ -2240,27 +2224,21 @@ public class CodegenModelGenerator extends AbstractCodegenModelGenerator {
       // Get the corresponding Buffer
       final Buffer firstFound = this.srSDFEdgeBuffers.get(subBuffProperty);
       Buffer buffer = null;
-      if (firstFound instanceof TwinBuffer) {
+      if (firstFound instanceof DistributedBuffer) {
         String coreBlockName = "";
         if (f.getType().equals(SpecialType.FORK) || f.getType().equals(SpecialType.BROADCAST)) {
           coreBlockName = source.getPropertyStringValue("Operator");
         } else {
           coreBlockName = target.getPropertyStringValue("Operator");
         }
-        TwinBuffer twinBuffer = (TwinBuffer) firstFound;
-        SubBuffer original = (SubBuffer) twinBuffer.getOriginal();
-        EList<Buffer> twins = twinBuffer.getTwins();
-        SubBuffer originalContainer = (SubBuffer) original.getContainer();
-        if (originalContainer.getContainer().getName().equals(coreBlockName)) {
-          buffer = original;
-        } else {
-          for (Buffer bufferTwinChecker : twins) {
-            SubBuffer subBufferChecker = (SubBuffer) bufferTwinChecker;
-            SubBuffer twinContainer = (SubBuffer) subBufferChecker.getContainer();
-            if (twinContainer.getContainer().getName().equals(coreBlockName)) {
-              buffer = subBufferChecker;
-              break;
-            }
+        DistributedBuffer distributedBuffer = (DistributedBuffer) firstFound;
+        EList<Buffer> repeatedBuffers = distributedBuffer.getDistributedCopies();
+        for (Buffer bufferRepeatedChecker : repeatedBuffers) {
+          SubBuffer subBufferChecker = (SubBuffer) bufferRepeatedChecker;
+          SubBuffer repeatedContainer = (SubBuffer) subBufferChecker.getContainer();
+          if (repeatedContainer.getContainer().getName().equals(coreBlockName)) {
+            buffer = subBufferChecker;
+            break;
           }
         }
       } else {
@@ -2367,22 +2345,16 @@ public class CodegenModelGenerator extends AbstractCodegenModelGenerator {
     final BufferProperties lastBuffProperty = bufferAggregate.get(0);
     final Buffer lastBufferFirstFound = this.srSDFEdgeBuffers.get(lastBuffProperty);
     Buffer lastBuffer = null;
-    if (lastBufferFirstFound instanceof TwinBuffer) {
+    if (lastBufferFirstFound instanceof DistributedBuffer) {
       String coreBlockName = operatorBlock.getName();
-      TwinBuffer twinBuffer = (TwinBuffer) lastBufferFirstFound;
-      SubBuffer original = (SubBuffer) twinBuffer.getOriginal();
-      EList<Buffer> twins = twinBuffer.getTwins();
-      SubBuffer originalContainer = (SubBuffer) original.getContainer();
-      if (originalContainer.getContainer().getName().equals(coreBlockName)) {
-        lastBuffer = original;
-      } else {
-        for (Buffer bufferTwinChecker : twins) {
-          SubBuffer subBufferChecker = (SubBuffer) bufferTwinChecker;
-          SubBuffer twinContainer = (SubBuffer) subBufferChecker.getContainer();
-          if (twinContainer.getContainer().getName().equals(coreBlockName)) {
-            lastBuffer = subBufferChecker;
-            break;
-          }
+      DistributedBuffer distributedBuffer = (DistributedBuffer) lastBufferFirstFound;
+      EList<Buffer> repeatedBuffers = distributedBuffer.getDistributedCopies();
+      for (Buffer bufferRepeatedChecker : repeatedBuffers) {
+        SubBuffer subBufferChecker = (SubBuffer) bufferRepeatedChecker;
+        SubBuffer repeatedContainer = (SubBuffer) subBufferChecker.getContainer();
+        if (repeatedContainer.getContainer().getName().equals(coreBlockName)) {
+          lastBuffer = subBufferChecker;
+          break;
         }
       }
     } else {
@@ -2500,9 +2472,9 @@ public class CodegenModelGenerator extends AbstractCodegenModelGenerator {
   }
 
   /**
-   * This method create a {@link TwinBuffer} for each {@link SubBuffer}. {@link SubBuffer} information are retrieved
-   * from the {@link #megs} of the {@link CodegenModelGenerator} . All created {@link SubBuffer} are referenced in the
-   * {@link #srSDFEdgeBuffers} map.
+   * This method create a {@link DistributedBuffer} for each {@link SubBuffer}. {@link SubBuffer} information are
+   * retrieved from the {@link #megs} of the {@link CodegenModelGenerator} . All created {@link SubBuffer} are
+   * referenced in the {@link #srSDFEdgeBuffers} map.
    *
    * @param parentBuffer
    *          the {@link Buffer} containing the generated {@link SubBuffer}
@@ -2511,7 +2483,7 @@ public class CodegenModelGenerator extends AbstractCodegenModelGenerator {
    * @return the total size of the subbuffers
    *
    */
-  protected long generateTwinSubBuffers(final Buffer twinParentBuffer, final DAGEdge dagEdge) {
+  protected long generateTwinSubBuffers(final Buffer repeatedParentBuffer, final DAGEdge dagEdge) {
 
     /*
      * Backup original
@@ -2524,9 +2496,9 @@ public class CodegenModelGenerator extends AbstractCodegenModelGenerator {
       backUpOriginal.put(subBufferProperties, buff);
     }
     /*
-     * Generate subbuffers for the twin buffer
+     * Generate subbuffers for the distributed buffer
      */
-    final long aggregateOffset = generateSubBuffers(twinParentBuffer, dagEdge);
+    final long aggregateOffset = generateSubBuffers(repeatedParentBuffer, dagEdge);
 
     /*
      * Get new buffers
@@ -2537,10 +2509,10 @@ public class CodegenModelGenerator extends AbstractCodegenModelGenerator {
       newBuffers.put(subBufferProperties, buff);
     }
     /*
-     * Generate and store twin buffers
+     * Generate and store repeated buffers
      */
     for (final BufferProperties subBufferProperties : buffers) {
-      TwinBuffer duplicatedBuffer = generateTwinBuffer(backUpOriginal.get(subBufferProperties),
+      DistributedBuffer duplicatedBuffer = generateDistributedBuffer(backUpOriginal.get(subBufferProperties),
           newBuffers.get(subBufferProperties));
       this.srSDFEdgeBuffers.put(subBufferProperties, duplicatedBuffer);
     }
@@ -2548,25 +2520,24 @@ public class CodegenModelGenerator extends AbstractCodegenModelGenerator {
   }
 
   /**
-   * This method create a {@link TwinBuffer} aggregated in the given original {@link Buffer}.
+   * This method create a {@link DistributedBuffer} aggregated in the given original {@link Buffer}.
    *
    * @param originalBuffer
    *          the {@link Buffer} containing the originalBuffer
-   * @param twinBuffer
-   *          the {@link Buffer} twinBuffer
-   * @return the twin buffer
+   * @param repeatedBuffer
+   *          the {@link Buffer} repeatedBuffer
+   * @return the distributed buffer
    *
    */
-  protected TwinBuffer generateTwinBuffer(final Buffer originalBuffer, final Buffer twinBuffer) {
+  protected DistributedBuffer generateDistributedBuffer(final Buffer originalBuffer, final Buffer repeatedBuffer) {
 
-    TwinBuffer duplicatedBuffer = CodegenFactory.eINSTANCE.createTwinBuffer();
-    if (originalBuffer instanceof TwinBuffer) {
-      duplicatedBuffer.setOriginal(((TwinBuffer) originalBuffer).getOriginal());
-      duplicatedBuffer.getTwins().addAll(((TwinBuffer) originalBuffer).getTwins());
+    DistributedBuffer duplicatedBuffer = CodegenFactory.eINSTANCE.createDistributedBuffer();
+    if (originalBuffer instanceof DistributedBuffer) {
+      duplicatedBuffer.getDistributedCopies().addAll(((DistributedBuffer) originalBuffer).getDistributedCopies());
     } else {
-      duplicatedBuffer.setOriginal(originalBuffer);
+      duplicatedBuffer.getDistributedCopies().add(originalBuffer);
     }
-    duplicatedBuffer.getTwins().add(twinBuffer);
+    duplicatedBuffer.getDistributedCopies().add(repeatedBuffer);
     return duplicatedBuffer;
   }